+++ conf/script/src/build_system/cmd/setup/cli/compiler_shell_env.py
